refactor: IOSSwitch component 분리
